requirements.txt
fastapi
uvicorn
ddgs
openai
python-dotenv
